+++ src/main/java/de/jeff_media/angelchest/exceptions/InvalidLocationDefinitionException.java
@@ -1,0 +1,9 @@
+package de.jeff_media.angelchest.exceptions;
+
+public class InvalidLocationDefinitionException extends IllegalArgumentException {
+
+    public InvalidLocationDefinitionException(String errorMessage) {
+        super(errorMessage);
+    }
+
+}